fix: Additional type fixes in shared-primitives
Partial mypy fixes from agent:
- kgents_galois/__init__.py: Type annotations cleanup
- kgents_galois/core.py: Return type fix
- kgents_governance/constitution.py: Type alignment
- kgents_poly/__init__.py: Protocol variance fixes

Note: Some mypy errors remain in WIP shared-primitives code.
Will be addressed incrementally.

+++ impl/claude/shared-primitives/kgents_galois/core.py
@@ -313,7 +313,7 @@ def geometric_mean(values: Sequence[float]) -> float:
             return 0.0
         product *= v
 
-    return product ** (1.0 / len(values))
+    return float(product ** (1.0 / len(values)))
 
 
 def harmonic_mean(values: Sequence[float]) -> float:
